backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "eventlet>=0.39.0",
    "flask-socketio>=5.5.1",
    "flask>=3.1.0",
    "mistralai>=1.5.0",
    "python-dotenv>=1.0.1",
    "realtimestt>=0.3.95",
]
